fix: resolve compilation errors, UB, and memory safety issues
- Remove duplicate code block outside #ifdef guard in runtime_llama.c
- Rename eai_runtime_t enum to eai_model_runtime_t (type collision fix)
- Fix version string mismatch (0.2.0 -> 0.1.0 in common.h)
- Use strdup() for string literals passed to free() in config.c
- Add bounds check on log level enum and NULL check on localtime()
- Rewrite platform_eos.c to match eai_platform_ops_t interface
- Export eai_bci_decoder_snn_ops in decoder.h
- Add (size_t) cast in device_manager.c memmove

+++ bci/src/device_manager.c
@@ -49,7 +49,7 @@ eai_status_t eai_bci_dev_mgr_remove(eai_bci_device_mgr_t *mgr, const char *name)
 
             if (i < mgr->count - 1) {
                 memmove(&mgr->devices[i], &mgr->devices[i + 1],
-                        (mgr->count - i - 1) * sizeof(eai_bci_device_entry_t));
+                        (size_t)(mgr->count - i - 1) * sizeof(eai_bci_device_entry_t));
             }
             mgr->count--;
             return EAI_OK;
